requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.2+cpu
torchvision==0.17.2+cpu

streamlit
streamlit-drawable-canvas
Pillow
psycopg2-binary
python-dotenv


